pyproject.toml
[project]
name = "cloudconduit"
version = "0.1.0"
description = "Unified API for connecting to Snowflake, Databricks, and S3 with credential management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "snowflake-connector-python[pandas]>=3.0.0",
    "databricks-sql-connector>=3.0.0",
    "boto3>=1.34.0",
    "pandas>=2.0.0",
    "keyring>=24.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]
authors = [
    {name = "CloudConduit Contributors"},
]
keywords = ["snowflake", "databricks", "s3", "aws", "data", "connector"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "myst-parser>=2.0.0",
]
all = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0", 
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0", 
    "sphinx-autodoc-typehints>=1.24.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/cloudconduit"
Repository = "https://github.com/yourusername/cloudconduit"
Issues = "https://github.com/yourusername/cloudconduit/issues"

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
